# hoppscotch/dflowp-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DFlowP API",
    "version": "1.0.0",
    "description": "OpenAPI-Definition fuer den Import in Hoppscotch."
  },
  "servers": [
    {
      "url": "https://dflowp.ddns.net/mcp",
      "description": "Produktive HTTPS-Schnittstelle"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/v1/datasets": {
      "get": {
        "summary": "Datasets auflisten",
        "description": "Liefert Datasets mit Pagination.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginierte Dataset-Liste",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Dataset anlegen",
        "description": "Entweder data_ids (bestehende Data-IDs) oder rows (neue Data-Zeilen); genau eine Variante.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Angelegtes Dataset-Dokument",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableError"
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "summary": "Dataset im Detail lesen",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dataset-Dokument",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/v1/data-items": {
      "get": {
        "summary": "DataItems auflisten (Data + Dataset)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginierte DataItem-Liste",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "summary": "Data-Dokument anlegen",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataItemCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Angelegtes Data-Dokument",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          }
        }
      }
    },
    "/api/v1/processes": {
      "get": {
        "summary": "Prozesse auflisten",
        "description": "Liefert Prozesse mit Pagination.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginierte Prozess-Liste",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}": {
      "get": {
        "summary": "Prozess im Detail lesen",
        "parameters": [
          {
            "name": "process_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prozess-Dokument",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/v1/subprocesses": {
      "get": {
        "summary": "Subprozesse auflisten",
        "description": "Liefert Subprozesse mit Pagination.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginierte Subprozess-Liste",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/subprocesses/{subprocess_id}": {
      "get": {
        "summary": "Subprozess im Detail lesen",
        "parameters": [
          {
            "name": "subprocess_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subprozess-Dokument",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDocument"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Muss dem Wert von DFlowP_API_Key entsprechen."
      }
    },
    "schemas": {
      "GenericDocument": {
        "type": "object",
        "additionalProperties": true
      },
      "DatasetCreateRequest": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Eindeutige Dataset-ID"
          },
          "data_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bestehende Data-IDs (doc_type=data)"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Neue Data-Zeilen; IDs werden als {id}_row_{i} vergeben"
          }
        }
      },
      "DataItemCreateRequest": {
        "type": "object",
        "required": ["content"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Optionale feste ID"
          },
          "content": {
            "type": "object",
            "additionalProperties": true
          },
          "type": {
            "type": "string",
            "default": "input"
          }
        }
      },
      "PaginatedListResponse": {
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenericDocument"
            }
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total_items": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Ungueltiger oder fehlender API-Key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Dokument nicht gefunden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ConflictError": {
        "description": "ID bereits vergeben.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnprocessableError": {
        "description": "Validierungsfehler (z. B. data_ids und rows gleichzeitig).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}